Implemented table filters

+++ backend/controllers/expense.js
@@ -1,9 +1,12 @@
+const asyncHandler = require("express-async-handler");
 const ExpenseSchema = require("../models/ExpenseModel");
+const UserSchema = require("../models/UserModel");
 
-exports.addExpense = async (req, res) => {
+exports.addExpense = asyncHandler(async (req, res) => {
   const { title, amount, category, description, date } = req.body;
 
   const income = ExpenseSchema({
+    user: req.user.id,
     title,
     amount,
     category,
@@ -11,35 +14,45 @@ exports.addExpense = async (req, res) => {
     date,
   });
 
-  try {
-    if (!title || !category || !description || !date) {
-      return res.status(400).json({ message: "All fields are required!" });
-    }
-    if (amount <= 0 || !amount === "number") {
-      return res
-        .status(400)
-        .json({ message: "Amount must be a positive number!" });
-    }
-    await income.save();
-    res.status(200).json({ message: "Expense Added" });
-  } catch (error) {
-    res.status(500).json({ message: "Server Error" });
+  if (!title || !category || !description || !date) {
+    res.status(400);
+    throw new Error("Please fill all fields");
   }
-
-  console.log(income);
-};
-
-exports.getExpense = async (req, res) => {
-  try {
-    const incomes = await ExpenseSchema.find().sort({ createdAt: -1 });
-    res.status(200).json(incomes);
-  } catch (error) {
-    res.status(500).json({ message: "Server Error" });
+  if (amount < 0 || !amount === "number") {
+    res.status(400);
+    throw new Error("Amount must be a positive number!");
   }
-};
+  const newExpense = await income.save();
+  res.status(200).json({ message: "Expense Added" });
+});
 
-exports.updateExpense = async (req, res) => {
+exports.getExpenses = asyncHandler(async (req, res) => {
+  const incomes = await ExpenseSchema.find({ user: req.user.id }).sort({
+    createdAt: -1,
+  });
+  res.status(200).json(incomes);
+});
+
+exports.updateExpense = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const income = await ExpenseSchema.findById(id);
+
+  if (!income) {
+    res.status(401);
+    throw new Error("Expense not found");
+  }
+
+  const user = await UserSchema.findById(req.user.id);
+
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  if (income.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Unauthorized");
+  }
   ExpenseSchema.findByIdAndUpdate(id, req.body, { new: true })
     .then((income) => {
       res.status(200).json({ message: "Expense updated successfully" });
@@ -47,15 +60,34 @@ exports.updateExpense = async (req, res) => {
     .catch((err) => {
       res.status(500).json({ message: "Server Error" });
     });
-};
+});
 
-exports.deleteExpense = async (req, res) => {
+exports.deleteExpense = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const income = await ExpenseSchema.findById(id);
+
+  if (!income) {
+    res.status(400);
+    throw new Error("Expense not found");
+  }
+
+  const user = await UserSchema.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (income.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   ExpenseSchema.findByIdAndDelete(id)
     .then((income) => {
-      res.status(200).json({ message: "Expense Deleted" });
+      res.status(200).json({ message: "Expense deleted successfully" });
     })
     .catch((err) => {
       res.status(500).json({ message: "Server Error" });
     });
-};
+});
