+++ backend/controllers/income.js
@@ -1,43 +1,57 @@
+const asyncHandler = require("express-async-handler");
 const IncomeSchema = require("../models/IncomeModel");
+const UserSchema = require("../models/UserModel");
 
-exports.addIncome = async (req, res) => {
-  const { title, amount, category, description, date } = req.body;
+exports.addIncome = asyncHandler(async (req, res) => {
+  const { title, amount, account, category, description, date } = req.body;
 
   const income = IncomeSchema({
+    user: req.user.id,
     title,
     amount,
+    account,
     category,
     description,
     date,
   });
 
-  try {
-    if (!title || !category || !description || !date) {
-      return res.status(400).json({ message: "Please fill all fields" });
-    }
-    if (amount < 0 || !amount === "number") {
-      return res
-        .status(400)
-        .json({ message: "Amount must be a positive number!" });
-    }
-    const newIncome = await income.save();
-    res.status(200).json({ message: "Income Added" });
-  } catch (err) {
-    res.status(500).json({ message: "Server Error" });
+  if ((!title || !amount || !account || !category || !description || !date)) {
+    res.status(400);
+    throw new Error("Please fill all fields");
   }
-};
-
-exports.getIncomes = async (req, res) => {
-  try {
-    const incomes = await IncomeSchema.find().sort({ createdAt: -1 });
-    res.status(200).json(incomes);
-  } catch (err) {
-    res.status(500).json({ message: "Server Error" });
+  if (amount < 0 || !amount === "number") {
+    res.status(400);
+    throw new Error("Amount must be a positive number!");
   }
-};
+  const newIncome = await income.save();
+  res.status(200).json({ message: "Income Added" });
+});
 
-exports.updateIncome = async (req, res) => {
+exports.getIncomes = asyncHandler(async (req, res) => {
+  const incomes = await IncomeSchema.find({ user: req.user.id }).sort({
+    createdAt: -1,
+  });
+  res.status(200).json(incomes);
+});
+
+exports.updateIncome = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const income = await IncomeSchema.findById(id);
+
+  if (!income) {
+    res.status(401);
+    throw new Error("Income not found");
+  }
+
+  if (!req.user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  if (income.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("Unauthorized");
+  }
   IncomeSchema.findByIdAndUpdate(id, req.body, { new: true })
     .then((income) => {
       res.status(200).json({ message: "Income updated successfully" });
@@ -45,10 +59,27 @@ exports.updateIncome = async (req, res) => {
     .catch((err) => {
       res.status(500).json({ message: "Server Error" });
     });
-};
+});
 
-exports.deleteIncome = async (req, res) => {
+exports.deleteIncome = asyncHandler(async (req, res) => {
   const { id } = req.params;
+  const income = await IncomeSchema.findById(id);
+
+  if (!income) {
+    res.status(400);
+    throw new Error("Income not found");
+  }
+
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (income.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   IncomeSchema.findByIdAndDelete(id)
     .then((income) => {
       res.status(200).json({ message: "Income deleted successfully" });
@@ -56,4 +87,4 @@ exports.deleteIncome = async (req, res) => {
     .catch((err) => {
       res.status(500).json({ message: "Server Error" });
     });
-};
+});
